Added validator for import completions form to prevent exceeding post request max length

+++ app/models/IndexListEntry.scala
@@ -1,3 +1,3 @@
 package models
 
-case class IndexListEntry(name: String, documents: Long, size: Long)
+case class IndexListEntry(name: String, documents: Long, size: Double)
